backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.functools-lru-cache==1.5
bcrypt==3.2.2
beautifulsoup4==4.8.1
cchardet==2.1.7
certifi==2018.11.29
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.3.7
colorclass==2.2.0
cryptography==37.0.4
Cython==0.29.30
decorator==5.1.1
Django==3.2.13
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
elasticsearch==7.9.1
frozenlist==1.3.0
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
importlib-resources==5.9.0
json2html==1.3.0
jsonpatch==1.32
jsonpath-rw==1.4.0
jsonpointer==2.3
jsonschema==4.6.0
lxml==4.9.0
multidict==6.0.2
ndg-httpsclient==0.5.1
nodeenv==1.6.0
npyscreen==4.10.5
numpy==1.17.3
#uncomment for latest supported version
#numpy>=1.23.0
paramiko==2.4.2
paramiko-expect==0.2.8
pathtools==0.1.2
pexpect==4.7.0
ply==3.11
prettytable==3.3.0
prompt-toolkit==2.0.10
protobuf==3.8.0
psutil==5.6.2
ptyprocess==0.7.0
pyasn1==0.4.8
pycodestyle==2.3.1
pycparser==2.21
PyJWT==2.4.0
#PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==19.0.0
pyproject-toml==0.0.10
pyrsistent==0.18.1
pytz==2022.1
PyYAML==5.1
recordtype==1.3
redfish==3.0.0
requests==2.22.0
#requests==2.32.3
robotbackgroundlogger==1.2
robotframework==3.1.2
robotframework-libdoc2json==0.4
robotframework-lint==0.9
robotframework-requests==0.5.0
robotframework-scplibrary==1.2.0
robotframework-sshlibrary==3.6.0
scp==0.13.0
setuptools==63.1.0
semantic-version==2.10.0
setuptools-rust==1.4.0
setuptools-svn==1.2
simplejson==3.16.0
six==1.16.0
slackclient==2.7.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
termcolor==1.1.0
terminaltables==3.1.0
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.25.3
#urllib3==2.2.3
urlparse2==1.1.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.56.0
whitenoise==6.2.0
wincertstore==0.2
xmljson==0.2.0
yarl==1.7.2
zipp==3.8.1